Multi_Modal_AI_Agent/requirements.txt
openai>=1.0.0
anthropic>=0.7.0
streamlit>=1.28.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
aiohttp>=3.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.0
markdown>=3.5.0
python-dotenv>=1.0.0
pydantic>=2.4.0
websockets>=11.0.0
modal>=0.57.0
semanticscholar>=0.10.0
arxiv>=2.1.0
crossrefapi>=1.0.0